pyproject.toml
[project]
name = "mp2i-discord-bot"
version = "0.1.0"
description = "Discord Bot for Prépas MP2I/MPI discord server"
readme = "README.md"
requires-python = ">=3.8,<=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "discord-py>=2.6.4",
    "humanize>=4.10.0",
    "psycopg>=3.2.12",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.44",
]
